airflow/requirements.txt
kafka-python-ng==2.2.3
pyarrow==16.1.0
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
Pillow==10.3.0
